Fixed division metod

+++ Kalkylator/Program.cs
@@ -77,7 +77,7 @@
                 return a * b;
                 }
 
-            static double Division(int a, int b)
+            static double Division(double a, double b)
             {
                 return a / b;
             }
